volcli: abstract command list / flags away in an easier-to-integrate way with contivctl
